+++ contrib/opencensus-ext-azure/opencensus/ext/azure/log_exporter/__init__.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 import logging
+import os
 import random
 import threading
 import time
@@ -63,10 +64,10 @@ class BaseLogHandler(logging.Handler):
         self._worker = Worker(self._queue, self)
         self._worker.start()
         # start statsbeat on exporter instantiation
-        if self.options.enable_stats_metrics:
-            statsbeat_metrics.collect_statsbeat_metrics(
-                self.options.instrumentation_key
-            )
+        if not os.environ.get("APPLICATIONINSIGHTS_STATSBEAT_DISABLED_ALL"):
+            statsbeat_metrics.collect_statsbeat_metrics(self.options)
+        # For redirects
+        self._consecutive_redirects = 0  # To prevent circular redirects
 
     def _export(self, batch, event=None):  # pragma: NO COVER
         try: