+++ contrib/opencensus-ext-azure/tests/test_azure_log_exporter.py
@@ -45,7 +45,6 @@ class CustomLogHandler(log_exporter.BaseLogHandler):
         self.callback = callback
         super(CustomLogHandler, self).__init__(
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
-            enable_stats_metrics=False,
         )
 
     def export(self, batch):
@@ -53,6 +52,15 @@ class CustomLogHandler(log_exporter.BaseLogHandler):
 
 
 class TestBaseLogHandler(unittest.TestCase):
+
+    def setUp(self):
+        os.environ["APPLICATIONINSIGHTS_STATSBEAT_DISABLED_ALL"] = "true"
+        return super(TestBaseLogHandler, self).setUp()
+
+    def tearDown(self):
+        del os.environ["APPLICATIONINSIGHTS_STATSBEAT_DISABLED_ALL"]
+        return super(TestBaseLogHandler, self).tearDown()
+
     def test_basic(self):
         logger = logging.getLogger(self.id())
         handler = CustomLogHandler(10, lambda batch: None)
@@ -74,14 +82,17 @@ class TestBaseLogHandler(unittest.TestCase):
 
 
 class TestAzureLogHandler(unittest.TestCase):
+
+    def setUp(self):
+        os.environ["APPLICATIONINSIGHTS_STATSBEAT_DISABLED_ALL"] = "true"
+        return super(TestAzureLogHandler, self).setUp()
+
+    def tearDown(self):
+        del os.environ["APPLICATIONINSIGHTS_STATSBEAT_DISABLED_ALL"]
+        return super(TestAzureLogHandler, self).tearDown()
+
     def test_ctor(self):
-        from opencensus.ext.azure.common import Options
-        instrumentation_key = Options._default.instrumentation_key
-        Options._default.instrumentation_key = None
-        self.assertRaises(ValueError, lambda: log_exporter.AzureLogHandler(
-            enable_stats_metrics=False,
-        ))
-        Options._default.instrumentation_key = instrumentation_key
+        self.assertRaises(ValueError, lambda: log_exporter.AzureLogHandler(connection_string="", instrumentation_key=""))  # noqa: E501
 
     def test_invalid_sampling_rate(self):
         with self.assertRaises(ValueError):
@@ -104,7 +115,6 @@ class TestAzureLogHandler(unittest.TestCase):
 
     def test_init_handler_with_queue_capacity(self):
         handler = log_exporter.AzureLogHandler(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             queue_capacity=500,
         )
@@ -123,7 +133,6 @@ class TestAzureLogHandler(unittest.TestCase):
     def test_exception(self, requests_mock):
         logger = logging.getLogger(self.id())
         handler = log_exporter.AzureLogHandler(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             storage_path=os.path.join(TEST_FOLDER, self.id()),
         )
@@ -141,7 +150,6 @@ class TestAzureLogHandler(unittest.TestCase):
     def test_exception_with_custom_properties(self, requests_mock):
         logger = logging.getLogger(self.id())
         handler = log_exporter.AzureLogHandler(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             storage_path=os.path.join(TEST_FOLDER, self.id()),
         )
@@ -167,7 +175,6 @@ class TestAzureLogHandler(unittest.TestCase):
     @mock.patch('requests.post', return_value=mock.Mock())
     def test_export_empty(self, request_mock):
         handler = log_exporter.AzureLogHandler(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             storage_path=os.path.join(TEST_FOLDER, self.id()),
         )
@@ -180,7 +187,6 @@ class TestAzureLogHandler(unittest.TestCase):
     def test_export_failure(self, log_record_to_envelope_mock):
         log_record_to_envelope_mock.return_value = ['bar']
         handler = log_exporter.AzureLogHandler(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             storage_path=os.path.join(TEST_FOLDER, self.id()),
         )
@@ -194,7 +200,6 @@ class TestAzureLogHandler(unittest.TestCase):
 
     def test_log_record_to_envelope(self):
         handler = log_exporter.AzureLogHandler(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             storage_path=os.path.join(TEST_FOLDER, self.id()),
         )
@@ -211,7 +216,6 @@ class TestAzureLogHandler(unittest.TestCase):
     def test_log_record_with_custom_properties(self, requests_mock):
         logger = logging.getLogger(self.id())
         handler = log_exporter.AzureLogHandler(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             storage_path=os.path.join(TEST_FOLDER, self.id()),
         )
@@ -233,7 +237,6 @@ class TestAzureLogHandler(unittest.TestCase):
     def test_log_with_invalid_custom_properties(self, requests_mock):
         logger = logging.getLogger(self.id())
         handler = log_exporter.AzureLogHandler(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             storage_path=os.path.join(TEST_FOLDER, self.id()),
         )
@@ -260,7 +263,6 @@ class TestAzureLogHandler(unittest.TestCase):
     def test_log_record_sampled(self, requests_mock):
         logger = logging.getLogger(self.id())
         handler = log_exporter.AzureLogHandler(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             logging_sampling_rate=1.0,
         )
@@ -280,7 +282,6 @@ class TestAzureLogHandler(unittest.TestCase):
     def test_log_record_not_sampled(self, requests_mock):
         logger = logging.getLogger(self.id())
         handler = log_exporter.AzureLogHandler(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             logging_sampling_rate=0.0,
         )
@@ -294,26 +295,26 @@ class TestAzureLogHandler(unittest.TestCase):
 
 
 class TestAzureEventHandler(unittest.TestCase):
+    def setUp(self):
+        os.environ["APPLICATIONINSIGHTS_STATSBEAT_DISABLED_ALL"] = "true"
+        return super(TestAzureEventHandler, self).setUp()
+
+    def tearDown(self):
+        del os.environ["APPLICATIONINSIGHTS_STATSBEAT_DISABLED_ALL"]
+        return super(TestAzureEventHandler, self).setUp()
+
     def test_ctor(self):
-        from opencensus.ext.azure.common import Options
-        instrumentation_key = Options._default.instrumentation_key
-        Options._default.instrumentation_key = None
-        self.assertRaises(ValueError, lambda: log_exporter.AzureEventHandler(
-            enable_stats_metrics=False,
-        ))
-        Options._default.instrumentation_key = instrumentation_key
+        self.assertRaises(ValueError, lambda: log_exporter.AzureEventHandler(connection_string="", instrumentation_key=""))  # noqa: E501
 
     def test_invalid_sampling_rate(self):
         with self.assertRaises(ValueError):
             log_exporter.AzureEventHandler(
-                enable_stats_metrics=False,
                 instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
                 logging_sampling_rate=4.0,
             )
 
     def test_init_handler_with_proxies(self):
         handler = log_exporter.AzureEventHandler(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             proxies='{"https":"https://test-proxy.com"}',
         )
@@ -325,7 +326,6 @@ class TestAzureEventHandler(unittest.TestCase):
 
     def test_init_handler_with_queue_capacity(self):
         handler = log_exporter.AzureEventHandler(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             queue_capacity=500,
         )
@@ -344,7 +344,6 @@ class TestAzureEventHandler(unittest.TestCase):
     def test_exception(self, requests_mock):
         logger = logging.getLogger(self.id())
         handler = log_exporter.AzureEventHandler(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             storage_path=os.path.join(TEST_FOLDER, self.id()),
         )
@@ -362,7 +361,6 @@ class TestAzureEventHandler(unittest.TestCase):
     def test_exception_with_custom_properties(self, requests_mock):
         logger = logging.getLogger(self.id())
         handler = log_exporter.AzureEventHandler(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             storage_path=os.path.join(TEST_FOLDER, self.id()),
         )
@@ -388,7 +386,6 @@ class TestAzureEventHandler(unittest.TestCase):
     @mock.patch('requests.post', return_value=mock.Mock())
     def test_export_empty(self, request_mock):
         handler = log_exporter.AzureEventHandler(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             storage_path=os.path.join(TEST_FOLDER, self.id()),
         )
@@ -401,7 +398,6 @@ class TestAzureEventHandler(unittest.TestCase):
     def test_export_failure(self, log_record_to_envelope_mock):
         log_record_to_envelope_mock.return_value = ['bar']
         handler = log_exporter.AzureEventHandler(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             storage_path=os.path.join(TEST_FOLDER, self.id()),
         )
@@ -415,7 +411,6 @@ class TestAzureEventHandler(unittest.TestCase):
 
     def test_log_record_to_envelope(self):
         handler = log_exporter.AzureEventHandler(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             storage_path=os.path.join(TEST_FOLDER, self.id()),
         )
@@ -432,7 +427,6 @@ class TestAzureEventHandler(unittest.TestCase):
     def test_log_record_with_custom_properties(self, requests_mock):
         logger = logging.getLogger(self.id())
         handler = log_exporter.AzureEventHandler(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             storage_path=os.path.join(TEST_FOLDER, self.id()),
         )
@@ -454,7 +448,6 @@ class TestAzureEventHandler(unittest.TestCase):
     def test_log_with_invalid_custom_properties(self, requests_mock):
         logger = logging.getLogger(self.id())
         handler = log_exporter.AzureEventHandler(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             storage_path=os.path.join(TEST_FOLDER, self.id()),
         )
@@ -481,7 +474,6 @@ class TestAzureEventHandler(unittest.TestCase):
     def test_log_record_sampled(self, requests_mock):
         logger = logging.getLogger(self.id())
         handler = log_exporter.AzureEventHandler(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             logging_sampling_rate=1.0,
         )
@@ -501,7 +493,6 @@ class TestAzureEventHandler(unittest.TestCase):
     def test_log_record_not_sampled(self, requests_mock):
         logger = logging.getLogger(self.id())
         handler = log_exporter.AzureEventHandler(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             logging_sampling_rate=0.0,
         )