+++ contrib/opencensus-ext-azure/tests/test_azure_metrics_exporter.py
@@ -12,13 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import os
 import unittest
 from datetime import datetime
 
 import mock
 
 from opencensus.common import utils
-from opencensus.ext.azure.common import Options
 from opencensus.ext.azure.common.protocol import DataPoint
 from opencensus.ext.azure.metrics_exporter import (
     MetricsExporter,
@@ -60,19 +60,22 @@ def create_metric():
 
 
 class TestAzureMetricsExporter(unittest.TestCase):
+
+    def setUp(self):
+        os.environ["APPLICATIONINSIGHTS_STATSBEAT_DISABLED_ALL"] = "true"
+        return super(TestAzureMetricsExporter, self).setUp()
+
+    def tearDown(self):
+        del os.environ["APPLICATIONINSIGHTS_STATSBEAT_DISABLED_ALL"]
+        return super(TestAzureMetricsExporter, self).tearDown()
+
     def test_constructor_missing_key(self):
-        instrumentation_key = Options._default.instrumentation_key
-        Options._default.instrumentation_key = None
-        self.assertRaises(ValueError, lambda: MetricsExporter(
-            enable_stats_metrics=False,
-        ))
-        Options._default.instrumentation_key = instrumentation_key
+        self.assertRaises(ValueError, lambda: MetricsExporter(connection_string="", instrumentation_key=""))  # noqa: E501
 
     def test_constructor_invalid_batch_size(self):
         self.assertRaises(
             ValueError,
             lambda: MetricsExporter(
-                enable_stats_metrics=False,
                 instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
                 max_batch_size=-1
             ))
@@ -81,7 +84,6 @@ class TestAzureMetricsExporter(unittest.TestCase):
     def test_export_metrics(self, requests_mock):
         metric = create_metric()
         exporter = MetricsExporter(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd')
         requests_mock.return_value.text = '{"itemsReceived":1,'\
                                           '"itemsAccepted":1,'\
@@ -97,7 +99,6 @@ class TestAzureMetricsExporter(unittest.TestCase):
     def test_export_metrics_histogram(self):
         metric = create_metric()
         exporter = MetricsExporter(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd')
         metric.descriptor._type = MetricDescriptorType.CUMULATIVE_DISTRIBUTION
 
@@ -106,7 +107,6 @@ class TestAzureMetricsExporter(unittest.TestCase):
     @mock.patch('requests.post', return_value=mock.Mock())
     def test_export_metrics_empty(self, requests_mock):
         exporter = MetricsExporter(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd')
         exporter.export_metrics([])
 
@@ -116,7 +116,6 @@ class TestAzureMetricsExporter(unittest.TestCase):
     def test_export_metrics_full_batch(self, requests_mock):
         metric = create_metric()
         exporter = MetricsExporter(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             max_batch_size=1)
         requests_mock.return_value.status_code = 200
@@ -133,7 +132,6 @@ class TestAzureMetricsExporter(unittest.TestCase):
     def test_create_data_points(self):
         metric = create_metric()
         exporter = MetricsExporter(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd'
         )
         data_points = exporter._create_data_points(metric.time_series[0],
@@ -149,7 +147,6 @@ class TestAzureMetricsExporter(unittest.TestCase):
     def test_create_properties(self):
         metric = create_metric()
         exporter = MetricsExporter(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd'
         )
         properties = exporter._create_properties(metric.time_series[0],
@@ -161,7 +158,6 @@ class TestAzureMetricsExporter(unittest.TestCase):
     def test_create_properties_none(self):
         metric = create_metric()
         exporter = MetricsExporter(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd'
         )
         metric.time_series[0].label_values[0]._value = None
@@ -174,7 +170,6 @@ class TestAzureMetricsExporter(unittest.TestCase):
     def test_create_envelope(self):
         metric = create_metric()
         exporter = MetricsExporter(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd'
         )
         value = metric.time_series[0].points[0].value.value
@@ -203,7 +198,6 @@ class TestAzureMetricsExporter(unittest.TestCase):
         mock_thread = mock.Mock()
         mock_storage = mock.Mock()
         exporter = MetricsExporter(
-            enable_stats_metrics=False,
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd'
         )
         exporter.exporter_thread = mock_thread