+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -2,14 +2,21 @@
 
 ## Unreleased
 
+## 1.1.0
+Released 2021-10-05
+
 - Enable AAD authorization via TokenCredential
 ([#1021](https://github.com/census-instrumentation/opencensus-python/pull/1021))
-- Implement attach rate metrics via Statbeat
+- Implement attach rate metrics via Statsbeat
 ([#1053](https://github.com/census-instrumentation/opencensus-python/pull/1053))
-- Implement network metrics via Statbeat - Success count
+- Implement network metrics via Statsbeat - Success count
 ([#1059](https://github.com/census-instrumentation/opencensus-python/pull/1059))
-- Implement network metrics via Statbeat - Others
+- Implement network metrics via Statsbeat - Others
 ([#1062](https://github.com/census-instrumentation/opencensus-python/pull/1062))
+- Implement feature and instrumentation metrics via Statsbeat
+([#1076](https://github.com/census-instrumentation/opencensus-python/pull/1076))
+- Support stamp specific redirect in exporters
+([#1078](https://github.com/census-instrumentation/opencensus-python/pull/1078))
 
 ## 1.0.8
 Released 2021-05-13